Create Technical.md
For this project to really get somewhere, all the little the technical details have to be written down somewhere for everyone to understand, instead of working off lists of features. This document describes an implementation similar to urho-osp, with some text taken from its wiki, and some fresh. It is still incomplete, and also needs lots of additions and review to make sure it makes sense.